refactor: Move chain's utils to UtilService
