+++ apps/canonical-bridge-server/src/module/token/token.processor.ts
@@ -13,6 +13,7 @@ import { ITokenJob } from '@/module/token/token.interface';
 import { Web3Service } from '@/shared/web3/web3.service';
 import { TokenService } from '@/module/token/token.service';
 import { CACHE_MANAGER, Cache } from '@nestjs/cache-manager';
+import { UtilService } from '@/shared/util/util.service';
 
 @Processor(Queues.SyncToken)
 export class TokenProcessor extends WorkerHost {
@@ -21,6 +22,7 @@ export class TokenProcessor extends WorkerHost {
   constructor(
     private web3Service: Web3Service,
     private tokenService: TokenService,
+    private utilService: UtilService,
     @Inject(CACHE_MANAGER) private cache: Cache,
     @InjectQueue(Queues.SyncToken) private syncToken: Queue<ITokenJob>,
   ) {
@@ -105,7 +107,7 @@ export class TokenProcessor extends WorkerHost {
     const now = Date.now();
     const config = tokens
       .map((t) => {
-        const chainConfig = this.tokenService.getChainConfigByLlamaPlatform(t.platform);
+        const chainConfig = this.utilService.getChainConfigByLlamaPlatform(t.platform);
         return {
           ...t,
           chainId: chainConfig?.id ?? t.chainId,
@@ -117,7 +119,7 @@ export class TokenProcessor extends WorkerHost {
       .reduce((r, c) => {
         const { address, chainId } = c;
 
-        const formattedAddr = this.tokenService.getFormattedAddress(chainId, address);
+        const formattedAddr = this.utilService.getFormattedAddress(chainId, address);
         const key = formattedAddr ? `${chainId}:${formattedAddr}` : chainId;
 
         r[key] = { price: c.price, decimals: c.decimals };
@@ -136,8 +138,8 @@ export class TokenProcessor extends WorkerHost {
     const config = tokens
       .map((t) => {
         const chainConfig =
-          this.tokenService.getChainConfigByCmcPlatform(t.platformSlug) ||
-          this.tokenService.getChainConfigByCmcPlatform(t.slug);
+          this.utilService.getChainConfigByCmcPlatform(t.platformSlug) ||
+          this.utilService.getChainConfigByCmcPlatform(t.slug);
 
         return {
           ...t,
@@ -148,7 +150,7 @@ export class TokenProcessor extends WorkerHost {
       .reduce((r, c) => {
         const { address, chainId } = c;
 
-        const formattedAddr = this.tokenService.getFormattedAddress(chainId, address);
+        const formattedAddr = this.utilService.getFormattedAddress(chainId, address);
         const key = formattedAddr ? `${chainId}:${formattedAddr}` : chainId;
 
         r[key] = { price: c.price, id: c.id };
